requirements.txt
torch
pandas
openpyxl
scikit-learn
gensim
langchain
python-dotenv
google-generativeai
langchain-google-genai
langchain-openai
langchain_huggingface
langchain-community
pandas
faiss-cpu
langgraph
SpeechRecognition
pyttsx3
gTTS
playsound
symbl
sounddevice
vosk
streamlit
fastapi
uvicorn
langchain-experimental
tabulate
selenium
beautifulsoup4
webdriver-manager
lxml